fix: additional error validation and logging

+++ .local/bin/battery_notifier.sh
@@ -7,15 +7,31 @@ if [ -e "$lock_file" ]; then
 	exit 1
 fi
 
-touch "$lock_file"
+if ! touch "$lock_file"; then
+	echo "Failed to create lock file."
+	exit 1
+fi
 
 trap 'rm -f "$lock_file"' EXIT
 
 while true; do
-	battery_level=$(acpi -b | grep -P -o '[0-9]+(?=%)')
-	charging_status=$(acpi -b | grep -o "Charging")
+	if ! battery_info=$(acpi -b); then
+		echo "Failed to get battery information."
+		sleep 300
+		continue
+	fi
 
-	if [ "$battery_level" -le 20 ] && [ -z "$charging_status" ]; then
+    battery_level=$(echo "$battery_info" | grep 'Battery 0' | grep -P -o '[0-9]+(?=%)')
+
+	if ! [[ "$battery_level" =~ ^[0-9]+$ ]]; then
+		echo "Invalid battery level: $battery_level"
+		sleep 300
+		continue
+	fi
+
+    charging_status=$(echo "$battery_info" | grep 'Battery 0' | grep -o "Charging")
+
+	if [ "${battery_level}" -le 20 ] && [ -z "${charging_status}" ]; then
 		notify-send "Low Battery" "Battery level is at ${battery_level}%" -u critical
 	fi
